requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
python-multipart
pillow
numpy
